Correct the config when a bond and vlan are used together
When a bond and a vlan are used together, and the subnet assigned to
the vlan is not set to dhcp mode, there is no IP info for the
bond interface itself. It is gathered and attached on the vlan
interface. This checks that the bond is actually assigned to a subnet
before trying to write out static IP information.